refactor backendstorage for reuse
